requirements.txt
torch>=2.1.0
torchvision>=0.16.1
transformers>=4.43.0
fastapi>=0.111.0
uvicorn>=0.30.1
pandas>=2.2.2
numpy>=1.26.4
Pillow>=11.0.0,<12.0
